% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetr.R
\name{distanceFromSeparator}
\alias{distanceFromSeparator}
\title{Calculate signed distance of point from linear separator}
\usage{
distanceFromSeparator(data, weight)
}
\arguments{
\item{data}{matrix or vector of numeric values}

\item{weight}{vector of numeric values}
}
\value{
If both arguments are vectors of same length, it will return the inner product. If one argument is a vector, it will be promoted to either a row or column matrix.
}
\description{
Calculate signed distance of point from linear separator
}
\examples{
vec1 <- c(1, 2, 3)
vec2 <- c(4, 5, 6)
distanceFromSeparator(vec1, vec2)
}
