% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetr.R
\name{euclidNorm}
\alias{euclidNorm}
\title{Calculate euclidean norm of a numerical vector.}
\usage{
euclidNorm(x)
}
\arguments{
\item{x}{vector of numeric values.}
}
\value{
the square root of the sum of squared values in \code{x}.
}
\description{
Calculate euclidean norm of a numerical vector.
}
\examples{
euclidNorm(c(1, 3, 0, 4))
euclidNorm(c(-1, 0, 5))
}
