% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetr.R
\name{plot.perceptron}
\alias{plot.perceptron}
\title{Plot of a fitted perceptron model.}
\usage{
\method{plot}{perceptron}(x, ...)
}
\arguments{
\item{x}{an object of class \code{perceptron} as returned by perceptron.}

\item{...}{arguments passed to or from other methods}
}
\description{
Plot of a fitted perceptron model.
}
\examples{
data(iris)
head(iris, n=20)
irisSub <- iris[1:100, c(1, 3, 5)]
names(irisSub) <- c("sepal", "petal", "species")
head(irisSub)
formula <- formula(species ~ sepal + petal)
p1 <- newPerceptronModel(formula, irisSub)
plot(p1, title = "Perceptron Classifier")
}
