% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetr.R
\name{predict.perceptron}
\alias{predict.perceptron}
\title{Predict function for a fitted perceptron model.}
\usage{
\method{predict}{perceptron}(object, newdata, ...)
}
\arguments{
\item{object}{fitted \code{perceptron} model.}

\item{newdata}{data frame from values for which to predict the class}

\item{...}{arguments passed to or from other methods}
}
\description{
Predict function for a fitted perceptron model.
}
\examples{
data(iris)
irisSub <- iris[(1:100), c(1, 3, 5)]
names(irisSub) <- c("sepal", "petal", "species")
formula <- formula(species ~ sepal + petal)
p1 <- newPerceptronModel(formula, irisSub)
irisShuffled <- irisSub[sample(nrow(irisSub)),]
irisTraining <- irisShuffled[1:70,]
irisHoldout <- irisShuffled[70:100,]
holdOutX <- irisHoldout[, 1:2]
holdOutY <- irisHoldout[, 3]
holdOutY <- factor(holdOutY)
prediction <- predict(p1, holdOutX)
}
