% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetr.R
\name{signum}
\alias{signum}
\title{Determine sign of distance from linear separator}
\usage{
signum(data, weight)
}
\arguments{
\item{data}{matrix or data frame of numeric values.}

\item{weight}{vector of numeric values.}
}
\value{
-1 if distance is negative else 1
}
\description{
Determine sign of distance from linear separator
}
\examples{
vals <- data.frame(4, 3, -2, 9)
weights <- c(1, 1, 1, 1)
signum(vals, weights)
}
