% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetr.R
\name{summary.perceptron}
\alias{summary.perceptron}
\title{Assemble summary output of fitted perceptron model.}
\usage{
\method{summary}{perceptron}(object, ...)
}
\arguments{
\item{object}{an object of class \code{perceptron} as returned by perceptron.}

\item{...}{arguments passed to or from other methods}
}
\value{
\describe{
\item{coefficients}{vector of weight values normalized by euclidean distance}
\item{options}{list of character representation of modelling options}
\item{input}{number of input layer nodes}
\item{hidden}{number of hidden layer nodes (fixed to 0)}
\item{output}{number of output layer nodes (fixed to 1)}
}
}
\description{
Assemble summary output of fitted perceptron model.
}
\examples{
data(iris)
head(iris, n=20)
iris_sub <- iris[1:100, c(1, 3, 5)]
names(iris_sub) <- c("sepal", "petal", "species")
head(iris_sub)
formula <- formula(species ~ sepal + petal)
p1 <- newPerceptronModel(formula, iris_sub)
summary(p1)
}
