% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetr.R
\name{print.perceptron}
\alias{print.perceptron}
\title{Print perceptron model.}
\usage{
\method{print}{perceptron}(x, ...)
}
\arguments{
\item{x}{an object of class \code{perceptron} as returned by perceptron.}

\item{...}{arguments passed to or from other methods.}
}
\description{
Print perceptron model.
}
\examples{
data(iris)
head(iris, n=20)
iris_sub <- iris[1:100, c(1, 3, 5)]
names(iris_sub) <- c("sepal", "petal", "species")
head(iris_sub)
formula <- formula(species ~ sepal + petal)
p1 <- newPerceptronModel(formula, iris_sub)
print(p1)
}
